+++ lib/core/controller/register.ts
@@ -45,7 +45,16 @@ export function Register(controllerJs: string): ControllerBundle {
         if (new ControllerValidation().Validate(output))
             exp = output;
         else throw new Error();
+
+        Object.keys(exp).forEach(action => {
+            if (['System', 'Console', 'Runtime','View'].indexOf(action) !== -1)
+                throw new Error('Warning! Reserved action name detected. Please read the document and try again.')
+            else if (action.match(/[a-z\d_]+/) === null)
+                throw new Error('Warning! Invalid characters detected in action names. Check and retry!')
+        })
     } catch (error) {
+        if ((error as Error).message.indexOf('Warning') > -1)
+            throw error;
         throw ErrorManager.RenderError(CompileTimeError.SH010103);
     }