+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Developing
 
+### 0.0.7 March 15, 2018
+
+- Wrap ServerResponse to avoid accidently crash ServerHub.
+- Fixed multiple document mistakes.
+
 ### 0.0.6 March 14, 2018
 
 #### Add